Tweaked EnumDrinkType, worked on Ghasts, etc.
Added several return statements on EnumDrinkType, and tried to finish
debugging GhastHandler. Still very buggy, seeing as it sometimes kills
the Ghasts and gives you a glass bottle. Also sometimes results in
out-of-bounds errors on the Ghasts upon entering the world after. Also,
Bottled Ghast can now be ingested, which results in a glass bottle and
ghast tear. (Ghast tears can now be stacked to 16.)